Integrate into existing app structure
