+++ app/src/app/App.View.tsx
@@ -19,7 +19,7 @@ import 'react-toastify/dist/ReactToastify.css';
 import 'theme/overwrites.css';
 
 import { IonApp, IonPage, IonSplitPane } from '@ionic/react';
-import { DefaultAppProps } from 'app/Default/DefaultAppProps';
+import { DefaultAppProps } from 'app/AppProps';
 import GlobalLoading from 'components/GlobalLoading/GlobalLoading';
 import SideMenu from 'components/SideMenu/SideMenu';
 import React, { useEffect } from 'react';
@@ -29,23 +29,31 @@ import { getViewComponentToViewId } from 'utils/viewMetaInfoHandlers';
 
 const DefaultAppView = ({
   loadInitialState,
-  currentViewId,
+  currentViewID: currentViewId,
+  showSideMenu,
 }: DefaultAppProps): JSX.Element => {
   useEffect(loadInitialState, [loadInitialState]);
+  const page = (
+    <IonPage id="mainViewContentId">
+      {React.createElement(getViewComponentToViewId(currentViewId))}
+    </IonPage>
+  );
   return (
     <IonApp>
-      <IonSplitPane
-        when="lg"
-        contentId="mainViewContentId"
-        class="split-pane-fixed"
-      >
-        {/* Side Menu (mainViewContentId used here!) */}
-        <SideMenu contentId="mainViewContentId" />
-        {/* Page Content */}
-        <IonPage id="mainViewContentId">
-          {React.createElement(getViewComponentToViewId(currentViewId))}
-        </IonPage>
-      </IonSplitPane>
+      {showSideMenu ? (
+        <IonSplitPane
+          when="lg"
+          contentId="mainViewContentId"
+          class="split-pane-fixed"
+        >
+          {/* Side Menu (mainViewContentId used here!) */}
+          <SideMenu contentId="mainViewContentId" />
+          {/* Page Content */}
+          {page}
+        </IonSplitPane>
+      ) : (
+        page
+      )}
       <ReactTooltip className="tooltip-fixed" html={true} place={'bottom'} />
       <GlobalLoading />
       <ToastContainer
