+++ app/src/app/App.tsx
@@ -1,13 +1,28 @@
-import DataViewerApp from 'app/DataViewer/DataViewerApp';
-import DataViewerAppHost from 'app/DataViewer/Host/DataViewerAppHost';
-import { isDataviewerAppWindow } from 'app/DataViewer/Host/DataViewerAppHostContext';
-import DefaultApp from 'app/Default/DefaultApp';
-import React from 'react';
+import DefaultAppView from 'app/App.View';
+import { DefaultAppDispatchProps, DefaultAppStateProps } from 'app/AppProps';
+import { connect } from 'react-redux';
+import { getAlgorithms } from 'store/actions/AlgorithmsPageActions';
+import { SnowmanDispatch } from 'store/messages';
+import { Store } from 'store/models';
+import { ViewMetaInformationCollection } from 'structs/viewMetaInfoCollection';
 
-export default function App(): JSX.Element {
-  return (
-    <DataViewerAppHost>
-      {isDataviewerAppWindow() ? <DataViewerApp /> : <DefaultApp />}
-    </DataViewerAppHost>
-  );
-}
+const mapStateToProps = ({
+  RenderLogicStore: { currentViewID },
+}: Store): DefaultAppStateProps => ({
+  currentViewID,
+  showSideMenu: !ViewMetaInformationCollection.find(
+    ({ key }) => key === currentViewID
+  )?.hideSideMenu,
+});
+
+const mapDispatchToProps = (
+  dispatch: SnowmanDispatch
+): DefaultAppDispatchProps => ({
+  loadInitialState() {
+    dispatch(getAlgorithms()).then();
+  },
+});
+
+const DefaultApp = connect(mapStateToProps, mapDispatchToProps)(DefaultAppView);
+
+export default DefaultApp;
