--- app/src/app/DataViewer/Host/DataViewerAppHostProps.ts
@@ -1,5 +1,0 @@
-export type DataViewerAppHostDispatchProps = {
-  couldNotOpenChildWindow(): void;
-};
-
-export type DataViewerAppHostProps = DataViewerAppHostDispatchProps;
