+++ app/src/components/DataViewer/Table/TableBody.tsx
@@ -1,5 +1,5 @@
 import { TableContext } from 'components/DataViewer/Table/TableContext';
-import { tableHeaderHeight } from 'components/DataViewer/Table/TableHeader';
+import { tableHeaderHeight } from 'components/DataViewer/Table/TableHeader.View';
 import { TableBodyProps } from 'components/DataViewer/Table/TableProps';
 import TableRow from 'components/DataViewer/Table/TableRow';
 import React, { useContext } from 'react';
