+++ app/src/components/DataViewer/Table/TableProps.ts
@@ -1,6 +1,7 @@
 import { MutableRefObject } from 'react';
 import { Column, ColumnInstance, HeaderGroup } from 'react-table';
 import { IndexRange } from 'react-virtualized';
+import { SnowmanThunkAction } from 'store/messages';
 
 export type TableProps = {
   columns: string[];
@@ -8,7 +9,7 @@ export type TableProps = {
   onRowsRendered: (params: IndexRange) => void;
   rowsChanged: unknown;
   columnsChanged: unknown;
-  openDataViewerWindow: () => void;
+  openDataViewerWindow: () => SnowmanThunkAction<void>;
 };
 
 export type TableContentProps = {
@@ -17,18 +18,24 @@ export type TableContentProps = {
   onRowsRendered: (params: IndexRange) => void;
   width: number;
   resetTable: MutableRefObject<boolean>;
-  openDataViewerWindow: () => void;
+  openDataViewerWindow: () => SnowmanThunkAction<void>;
 };
 
 export type TableBodyProps = {
   onRowsRendered: (params: IndexRange) => void;
 };
 
-export type TableHeaderProps = {
+export type TableHeaderOwnProps = {
   headerGroups: HeaderGroup<string[]>[];
   visibleColumns: ColumnInstance<string[]>[];
   setColumnOrder: (
     updater: string[] | ((columnOrder: string[]) => string[])
   ) => void;
-  openDataViewerWindow: () => void;
+  openDataViewerWindow: () => SnowmanThunkAction<void>;
 };
+
+export type TableHeaderDispatchProps = {
+  performOpenDataViewerWindow: () => void;
+};
+
+export type TableHeaderProps = TableHeaderOwnProps & TableHeaderDispatchProps;
