+++ app/src/pages/StandaloneDataViewerPage/StandaloneDataViewerStyles.css
