+++ app/src/store/actions/RenderStoreActions.ts
@@ -1,9 +1,24 @@
+import { DataViewerOwnProps } from 'components/DataViewer/DataViewerProps';
+import {
+  DataViewerAppToClientActionType,
+  postActionToClient,
+} from 'pages/StandaloneDataViewerPage/actionsToClient';
+import {
+  DataViewerAppToHostActionType,
+  onActionFromClient,
+} from 'pages/StandaloneDataViewerPage/actionsToHost';
 import { RenderStoreActionTypes as actionTypes } from 'store/actions/actionTypes';
+import { showToast } from 'store/actions/GlobalIndicatorActions';
+import { SnowmanThunkAction } from 'store/messages';
+import { COULD_NOT_OPEN_CHILD_WINDOW_ERROR } from 'structs/statusMessages';
+import { ToastType } from 'types/ToastTypes';
+import { viewIdQueryParam } from 'types/ViewIdQueryParam';
 import { ViewIDs } from 'types/ViewIDs';
 import {
   easyPrimitiveAction,
   easyPrimitiveActionReturn,
 } from 'utils/easyActionsFactory';
+import { unwrapError } from 'utils/requestHandler';
 
 export const navigateTo = (aTarget: ViewIDs): easyPrimitiveActionReturn =>
   easyPrimitiveAction({
@@ -17,3 +32,65 @@ export const navigateToNextPage = (): easyPrimitiveActionReturn =>
     // reducer ignores payload
     payload: false,
   });
+
+const windowNameBase = 'snowman-benchmark';
+let nextWindowId = 0;
+const nextWindowName = (): string => {
+  return `${windowNameBase}-${++nextWindowId}-${window.name}`;
+};
+
+const openViewInNewWindow = (
+  viewId: ViewIDs
+): SnowmanThunkAction<Window | null> => (dispatch) => {
+  const targetWindow = window.open(
+    `${window.location.href}?${viewIdQueryParam}=${viewId}`,
+    nextWindowName()
+  );
+  if (!targetWindow) {
+    dispatch(showToast(COULD_NOT_OPEN_CHILD_WINDOW_ERROR, ToastType.Error));
+  }
+  return targetWindow;
+};
+
+export const openStandaloneDataViewerWindow = ({
+  loadTuples,
+  ...props
+}: DataViewerOwnProps): SnowmanThunkAction<void> => (dispatch) => {
+  const targetWindow = dispatch(
+    openViewInNewWindow(ViewIDs.STANDALONE_DATA_VIEWER)
+  );
+  if (targetWindow) {
+    window.addEventListener('beforeunload', () =>
+      postActionToClient(targetWindow, {
+        type: DataViewerAppToClientActionType.WINDOW_CLOSED,
+        payload: undefined,
+      })
+    );
+    onActionFromClient(
+      targetWindow,
+      DataViewerAppToHostActionType.LOAD_TUPLES,
+      ({ id, params }) =>
+        loadTuples(...params)
+          .then((response) =>
+            postActionToClient(targetWindow, {
+              type: DataViewerAppToClientActionType.TUPLES_LOADER_RESPONSE,
+              payload: { response, id },
+            })
+          )
+          .catch(async (error) =>
+            postActionToClient(targetWindow, {
+              type: DataViewerAppToClientActionType.TUPLES_LOADER_FAILED,
+              payload: { id, error: await unwrapError(error) },
+            })
+          )
+    );
+    onActionFromClient(targetWindow, DataViewerAppToHostActionType.READY, () =>
+      postActionToClient(targetWindow, {
+        type: DataViewerAppToClientActionType.DATA_VIEWER_PARAMS,
+        payload: props,
+      })
+    );
+  } else {
+    dispatch(showToast(COULD_NOT_OPEN_CHILD_WINDOW_ERROR, ToastType.Error));
+  }
+};
