+++ app/src/structs/statusMessages.ts
@@ -23,5 +23,4 @@ export const SUCCESS_TO_UPDATE_EXPERIMENT =
 export const SUCCESS_TO_DELETE_DATASET = 'Successfully deleted the dataset!';
 export const SUCCESS_TO_DELETE_EXPERIMENT =
   'Successfully deleted the experiment!';
-export const COULD_NOT_OPEN_DATA_VIEWER_IN_CHILD_WINDOW_ERROR =
-  'Could not open data viewer in child window.';
+export const COULD_NOT_OPEN_CHILD_WINDOW_ERROR = 'Could not open child window.';
