+++ app/src/types/ViewIdQueryParam.ts
@@ -1,0 +1,1 @@
+export const viewIdQueryParam = 'viewid';
